Correctly resolve patchdir directory, relative to repo workdir
The patch directory passed to regenerate_patches
must be a relative path.
It is resolved against the workdir of the repo with the patch files.

This bug was detected while working on the paper_patch.rs example.
It is dependent on the process CWD (ie. only triggered if CWD != repo workdir).